requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.25